Implement login. Added token to all requests which require it

+++ online-supermarket/src/app/service/producer.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Producer } from '../model/producer.model';
 
 @Injectable({
@@ -10,22 +10,27 @@ export class ProducerService {
   constructor(private httpClient: HttpClient) { }
 
   getAllProducers() {
-    return this.httpClient.get<Producer[]>("http://localhost:8083/producers");
+    const token = localStorage.getItem('token');
+    return this.httpClient.get<Producer[]>("http://localhost:8083/producers", {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   addNewProducer(producer: Producer) {
-    return this.httpClient.post("http://localhost:8083/producers", producer);
+    const token = localStorage.getItem('token');
+    return this.httpClient.post("http://localhost:8083/producers", producer, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   deleteProducer(id: number) {
-    return this.httpClient.delete("http://localhost:8083/producers/" + id);
+    const token = localStorage.getItem('token');
+    return this.httpClient.delete("http://localhost:8083/producers/" + id, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   getProducerById(id: number) {
-    return this.httpClient.get<Producer>("http://localhost:8083/producers/" + id);
+    const token = localStorage.getItem('token');
+    return this.httpClient.get<Producer>("http://localhost:8083/producers/" + id, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   updateProducer(id: number, producer: Producer) {
-    return this.httpClient.put("http://localhost:8083/producers/" + id, producer);
+    const token = localStorage.getItem('token');
+    return this.httpClient.put("http://localhost:8083/producers/" + id, producer, {headers: new HttpHeaders().set('Authorization', token)});
   }
 }
