+++ online-supermarket/src/app/service/supplier.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Supplier } from '../model/supplier.model';
 
 @Injectable({
@@ -10,22 +10,27 @@ export class SupplierService {
   constructor(private httpClient: HttpClient) { }
 
   getAllSuppliers() {
-    return this.httpClient.get<Supplier[]>('http://localhost:8083/suppliers');
+    const token = localStorage.getItem('token');
+    return this.httpClient.get<Supplier[]>('http://localhost:8083/suppliers', {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   addSupplier(supplier: Supplier) {
-    return this.httpClient.post("http://localhost:8083/suppliers/", supplier);
+    const token = localStorage.getItem('token');
+    return this.httpClient.post("http://localhost:8083/suppliers/", supplier, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   deleteSupplier(id: number) {
-    return this.httpClient.delete("http://localhost:8083/suppliers/"+ id);
+    const token = localStorage.getItem('token');
+    return this.httpClient.delete("http://localhost:8083/suppliers/"+ id, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   getSupplierById(id: number) {
-    return this.httpClient.get<Supplier>("http://localhost:8083/suppliers/" + id);
+    const token = localStorage.getItem('token');
+    return this.httpClient.get<Supplier>("http://localhost:8083/suppliers/" + id, {headers: new HttpHeaders().set('Authorization', token)});
   }
 
   editSupplier(supplier: Supplier, id: number) {
-    return this.httpClient.put("http://localhost:8083/suppliers/" + id, supplier);
+    const token = localStorage.getItem('token');
+    return this.httpClient.put("http://localhost:8083/suppliers/" + id, supplier, {headers: new HttpHeaders().set('Authorization', token)});
   }
 }
